python/get_secret_from_secret_manager/google_cloud_functions/requirements.txt
Flask
google-cloud-secret-manager
